fix(#7251): split argument-binding tests out of eotest to satisfy qulice file-length check
