+++ eo-parser/src/test/java/org/eolang/parser/EoTest.java
@@ -751,157 +751,6 @@ final class EoTest {
     }
 
     @Test
-    void acceptsAllUnboundHorizontalArgs() {
-        MatcherAssert.assertThat(
-            "all-unbound is a valid mode per R-6.6.2 and must parse without errors",
-            EoTest.render("[] > main", "  foo a b c > x"),
-            XhtmlMatchers.hasXPath("/object[not(errors)]")
-        );
-    }
-
-    @Test
-    void acceptsAllBoundHorizontalArgs() {
-        MatcherAssert.assertThat(
-            "all-bound is a valid mode per R-6.6.2 and must parse without errors",
-            EoTest.render("[] > main", "  foo a:x b:y > z"),
-            XhtmlMatchers.hasXPath("/object[not(errors)]")
-        );
-    }
-
-    @Test
-    void reportsMixedBindings() {
-        MatcherAssert.assertThat(
-            "mixed bound and unbound args in the same group must surface R-6.6.2",
-            EoTest.render("[] > main", "  foo a:x b > z"),
-            XhtmlMatchers.hasXPath(
-                "/object/errors/error[contains(text(),'argument bindings must be all-or-nothing')]"
-            )
-        );
-    }
-
-    @Test
-    void acceptsReversedReceiverBareAndArgsBound() {
-        MatcherAssert.assertThat(
-            "the reversed receiver is exempt; the remaining args may be all-bound per R-6.6.2/R-6.6.3",
-            EoTest.render("[] > main", "  if. cond then:a else:b > z"),
-            XhtmlMatchers.hasXPath("/object[not(errors)]")
-        );
-    }
-
-    @Test
-    void rejectsBoundReversedReceiver() {
-        MatcherAssert.assertThat(
-            "a receiver carrying `:x` must surface R-6.6.3",
-            EoTest.render("[] > main", "  if. cond:x then else > z"),
-            XhtmlMatchers.hasXPath(
-                "/object/errors/error[contains(text(),'reversed-dispatch receiver cannot carry a binding')]"
-            )
-        );
-    }
-
-    @Test
-    void rejectsMixedBindingsAmongReversedArgs() {
-        MatcherAssert.assertThat(
-            "the rule still applies to args after the receiver in a reversed dispatch",
-            EoTest.render("[] > main", "  if. cond then:a else > z"),
-            XhtmlMatchers.hasXPath(
-                "/object/errors/error[contains(text(),'argument bindings must be all-or-nothing')]"
-            )
-        );
-    }
-
-    @Test
-    void emitsAsForVerticalBindingOnIdentifier() {
-        MatcherAssert.assertThat(
-            "a vertical-arg line ending with `:label` must emit @as on the line's outermost <o>",
-            EoTest.render("foo > main", "  bar:tag"),
-            XhtmlMatchers.hasXPath("/object/o[@name='main']/o[@base='bar' and @as='tag']")
-        );
-    }
-
-    @Test
-    void emitsAsForVerticalBindingWithSuffix() {
-        MatcherAssert.assertThat(
-            "the binding may combine with `> name`, both attaching to the same <o>",
-            EoTest.render("foo > main", "  bar:tag > out"),
-            XhtmlMatchers.hasXPath(
-                "/object/o[@name='main']/o[@name='out' and @base='bar' and @as='tag']"
-            )
-        );
-    }
-
-    @Test
-    void emitsNumericVerticalBindingAsAlpha() {
-        MatcherAssert.assertThat(
-            "a numeric `:N` binding must become @as='αN' (R-3.12.2)",
-            EoTest.render("foo > main", "  bar:0"),
-            XhtmlMatchers.hasXPath("/object/o[@name='main']/o[@base='bar' and @as='α0']")
-        );
-    }
-
-    @Test
-    void rejectsMixedVerticalBindings() {
-        MatcherAssert.assertThat(
-            "vertical args under a vapplication must follow R-6.6.2 — mixing bound and unbound is rejected",
-            EoTest.render("[] > main", "  foo > app", "    a:x", "    b"),
-            XhtmlMatchers.hasXPath(
-                "/object/errors/error[contains(text(),'argument bindings must be all-or-nothing')]"
-            )
-        );
-    }
-
-    @Test
-    void acceptsAllBoundVerticalArgs() {
-        MatcherAssert.assertThat(
-            "vertical args may all carry bindings — the uniform pattern is valid",
-            EoTest.render("[] > main", "  foo > app", "    a:x", "    b:y"),
-            XhtmlMatchers.hasXPath("/object[not(errors)]")
-        );
-    }
-
-    @Test
-    void rejectsMixedBindingsUnderChainedVerticalMethod() {
-        MatcherAssert.assertThat(
-            "args of the last link of a same-indent method chain must follow R-6.6.2 too",
-            EoTest.render("[] > main", "  foo > app", "  .bar", "  .baz", "    a", "    b:y"),
-            XhtmlMatchers.hasXPath(
-                "/object/errors/error[contains(text(),'argument bindings must be all-or-nothing')]"
-            )
-        );
-    }
-
-    @Test
-    void acceptsLoneBoundArgUnderVerticalMethod() {
-        MatcherAssert.assertThat(
-            "a single arg under a vmethod is a uniform group, so its binding is valid",
-            EoTest.render("[] > main", "  foo > app", "  .bar", "    a:x"),
-            XhtmlMatchers.hasXPath("/object[not(errors)]")
-        );
-    }
-
-    @Test
-    void rejectsBindingOnFormationChild() {
-        MatcherAssert.assertThat(
-            "a plain child of a formation cannot carry a binding per R-3.12.3",
-            EoTest.render("[] > main", "  bar:tag > x"),
-            XhtmlMatchers.hasXPath(
-                "/object/errors/error[contains(text(),'binding allowed only in argument position')]"
-            )
-        );
-    }
-
-    @Test
-    void rejectsBindingOnVerticalReceiver() {
-        MatcherAssert.assertThat(
-            "the receiver of a vertical reversed dispatch cannot carry a binding either",
-            EoTest.render("foo > main", "  if.", "    cond:x", "    then", "    other"),
-            XhtmlMatchers.hasXPath(
-                "/object/errors/error[contains(text(),'reversed-dispatch receiver cannot carry a binding')]"
-            )
-        );
-    }
-
-    @Test
     void acceptsInlineVoidsOnReversedDispatch() {
         MatcherAssert.assertThat(
             "a trailing-dot object with an inline void suffix must parse with no errors and desugar to the same φ-as-reversed-dispatch shape as its two-line form",
